main/pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = ""
authors = [
    {name = "Kairat Zhanadylov",email = "kairatzhanadylov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pytest-django (>=4.9.0,<5.0.0)",
    "pytest-factoryboy (>=2.7.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
